refactor: minor improvements to .exp, .versions, .tug and move .cowsay (#108)
* update .exp and .versions

* minor tug.ts improvements, move cowsay.ts to funnies

* move cowsay.ts
