+++ src/commands/support/tug.ts
@@ -10,7 +10,9 @@ export const tug: CommandDefinition = {
         title: 'FlyByWire Support | My Aircraft is Stuck!',
         description: makeLines([
             'Make sure you do not have the "NW STRG DISC" message on the upper ECAM display. ',
+            '',
             'If you do, please press "Shift" + "P" on your keyboard to disconnect the invisible pushback tug.',
+            '',
             'Please read our guide [here](https://docs.flybywiresim.com/fbw-a32nx/support/reported-issues/#nose-wheel-steering-locked-nw-strg-disc) for more information. ',
         ]), 
     })),
